fix(api): refactor mcp tools
